# Fix https://github.com/jsk-ros-pkg/jsk_recognition/issues/2860