close #2860 

`libsiftfast` has so many static global variables, so it is not thread-safe when called from multiple nodelets. This PR adds a global mutex to restrict the access to those variables to make it thread-safe.

Since the entire mutex is taken, if n imagesift nodelet are loaded, the time required to compute each feature is n times longer. But I think it is fine because the node is implemented with ConnectionBasedNodelet. If the user doesn't subscribe, the slow calculation doesn't occur.  Making the user code work is much important.

The clean method is to make [libsiftfast](https://github.com/jsk-ros-pkg/jsk_3rdparty/tree/master/3rdparty/libsiftfast) thread-safe, but I think we should write it from scratch or use OpenCV if we do so. 

You can try the bug occurs just reverting 365b2d00050e6ef9a853ad21a046ac59adf07308 and execute `test_two_imagesift.test`