@mqcmd196 2025-04-23
```
requests.exceptions.ConnectionError: HTTPConnectionPool(host='jenkins.jsk.imi.i.u-tokyo.ac.jp', port=8080): Max retries exceeded with url: /crumbIssuer/api/json (Caused by NewConnectionError('<urllib3.connection.HTTPConnection object at 0x7f22e150c5d0>: Failed to establish a new connection: [Errno -5] No address associated with hostname',))
```