@k-okada 2025-04-28
can you compare calculation speed in single thread with and whitout this PR? because libsiftfast is originally developed to pursue high speed processing by several optimization technique, so I'd like to know how this changes the performance.