@mqcmd196 2025-04-30
> calculation speed in single thread

What do you mean? Making the nodelet's num_worker_thread=1 or making the libsiftfast's OpenMP flag OFF?