@k-okada 2025-05-01
sorry 
x single thread
o single nodelet
I just want to know if adding global mutex lock affect comptation speed .